Patient Tracking System
My project was a patient tracking system. First, I started by creating a login form. Then I created a registration form. To go directly from the login form to the registration form, I created a JLabel and added a feature. I created a crossing system that I call the main bridge. When we press the relevant buttons in this system, the JFrame with the text inside is sent to us. I created a class to provide the connection and provided the connection through that class. I added the data from the registration form to this database. I set all the Frames to appear on the page. I used Regex and Generic in the registration form. To ensure passage on the main bridge, for example; The admin will be able to enter all of them, the doctor will be able to enter the patient chart and patient login, and the user will only be able to enter the patient login. To achieve this, I created the Who class and assigned its value in the combo box on the login form to the element of this class. Then, I used the database's selection feature and added the features I gave in the example to the relevant buttons. Using JsplitPane, I separated the button containing the tables and the patient entry button. I added a menubar and there are save and close options in the file folder of the menubar. Color selection allows you to change the color of the title. File selection makes you select a file. The file's location is printed in the text to the right of the title. I start by opening the checkbox and the delete, update, and show all data buttons. I created PatientLoginDB and saved the appropriate data from the patient login into this database.